ai-part/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.7.2
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
pydantic==2.5.0
pillow==10.1.0
cloudscraper==1.2.71
setuptools>=65.0.0